requirements.txt
torch
torchvision
numpy
matplotlib
scipy
Pillow
tqdm
scikit-image
opencv-python
pyyaml
ipykernel
torchsummary
graphviz
torchviz